// src/store/booking/types.ts
export interface SectionWithSeats extends SeatSectionBasic {
  selectedSeats?: SeatPosition[];
}

export type ConnectionDirection = 'there' | 'back';

export const CLEAR_BOOKING = 'CLEAR_BOOKING';
export const REPLACE_TARIFFS = 'REPLACE_TARIFFS';
export const SAVE_CONNECTION = 'SAVE_CONNECTION';
export const SELECT_ADDONS = 'SELECT_ADDONS';
export const SELECT_CLASS = 'SELECT_CLASS';
export const SELECT_CODE_DISCOUNT = 'SELECT_CODE_DISCOUNT';
export const SELECT_PERC_DISCOUNT = 'SELECT_PERC_DISCOUNT';
export const SELECT_ROUTE = 'SELECT_ROUTE';
export const SELECT_SEATS = 'SELECT_SEATS';
export const SET_PRICE = 'SET_PRICE';
export const UPSELL_ADDONS = 'UPSELL_ADDONS';

export interface PayloadCommons {
  direction: ConnectionDirection;
}

export interface SelectRoutePayload extends PayloadCommons {
  fromStationId: number;
  routeId: RouteId;
  tariffs: Tariff[];
  toStationId: number;
  type?: TicketType;
}
export interface ReplaceTariffsPayload extends PayloadCommons {
  tariffs: Tariff[];
}

export interface SelectClassPayload extends PayloadCommons {
  flexiType?: FlexiType;
  lineGroupCode?: string;
  lineNumber?: number;
  price: number;
  priceSource?: string;
  seatClass: SeatClass;
  sections: SectionWithSeats[];
  type: TicketType;
}
// ...
export interface SaveConnectionPayload extends Partial<ConnectionPayload> {
  ignoreFavorite?: boolean;
}